hotfix20022024

+++ CRM/Models/DTO/WorkLocationDTO.cs
@@ -5,6 +5,6 @@
         public int Id { get; set; }
         public string City { get; set; } = null!;
         public string State { get; set; } = null!;
-        public decimal? Commissoninpercentage { get; set; }
+        public int? Commissoninpercentage { get; set; }
     }
 }
